requirements.txt
asn1crypto==0.24.0
cachetools==3.0.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==2.1.3
django-cors-headers==2.4.0
django-gcloud-storage==0.3.1
django-storages==1.7.1
google-api-core==1.5.2
google-auth==1.6.0
google-cloud-core==0.28.1
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
gunicorn==19.9.0
idna==2.7
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.7
requests==2.20.1
rsa==4.0
six==1.11.0